requirements/prod.txt
# Everything needed in production

# Flask
Flask>=2.1.2
Werkzeug==2.1.2
click>=8.1.3

# Database
Flask-SQLAlchemy>=2.5.1
SQLAlchemy~=1.4.37
psycopg2-binary==2.9.3

# Migrations
Flask-Migrate==3.1.0

# Forms
Flask-WTF==1.0.1
WTForms==3.0.1
email-validator>=1.2.1

# Deployment
gevent==21.12.0
gunicorn>=20.1.0
supervisor==4.2.4

# Flask Static Digest
Flask-Static-Digest==0.2.1

# Auth
Flask-Login==0.6.1
Flask-Bcrypt==1.0.1
Flask-User==1.0.2.2
Speaklater==1.3

# Caching
Flask-Caching>=1.11.1

# Debug toolbar
Flask-DebugToolbar==0.13.1

# Environment variable parsing
environs==9.5.0

#
celery==5.2.7
redis==4.3.3
flask_table==0.5.0
PyYAML==6.0
flower==1.0.0
markdown==3.3.7
flask-socketio==5.1.0

#JWT
jwt==1.3.1
PyJWT==1.7.1
Flask-JWT-Simple==0.0.3

# DB-Schemata
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.15.0

# Docker
docker==6.1.3